Fixed crushes when deleting last item

+++ trunk/CuteTorrent/LTorrent/TorrentManager.cpp
@@ -79,7 +79,7 @@ TorrentManager::TorrentManager()
 		ses->start_dht();
 
 	ses->set_settings(s_settings);
-	////qDebug() << "TorrentManager: intialisation completed";
+	//qDebug() << "TorrentManager: intialisation completed";
 	
 	}
 	catch(std::exception ex)
@@ -96,7 +96,7 @@ void TorrentManager::initSession()
 	QStringList filter;
 	
 	filter <<"*.torrent";
-	////qDebug() << "TorrentManager::initSession: getting torrentfiles we added";
+	//qDebug() << "TorrentManager::initSession: getting torrentfiles we added";
 	QStringList torrentFiles=dir.entryList(filter);
 	QFile path_infohashFile("CT_DATA/path.resume");
 	if (path_infohashFile.open(QFile::ReadOnly))
@@ -121,7 +121,7 @@ void TorrentManager::initSession()
 		return;	
 			
 	}
-	////qDebug() << "TorrentManager::initSession: found " <<torrentFiles.count();
+	//qDebug() << "TorrentManager::initSession: found " <<torrentFiles.count();
 	error_code ec;
 	for (QStringList::iterator i=torrentFiles.begin();i!=torrentFiles.end();i++)
 	{
@@ -161,13 +161,11 @@ bool TorrentManager::AddTorrent(QString path, QString save_path,QMap<QString,int
 		error_code ec;
 		
 		
-		t = new torrent_info(path.toAscii().data(), ec);
+		t = new torrent_info(path.toUtf8().data(), ec);
 		
 		if (ec)
 		{
-			QMessageBox msg;
-			msg.setText(ec.message().c_str());
-			msg.exec();
+			QMessageBox::warning(0,"Добавление торрента",QString::fromUtf8(ec.message().c_str()));
 			return false;
 		}
 	
@@ -195,7 +193,7 @@ bool TorrentManager::AddTorrent(QString path, QString save_path,QMap<QString,int
 					filepriorities->push_back(filePriorities[storrage.file_path(*i).c_str()]);
 				else
 				{
-					////qDebug() << "not found " << storrage.file_path(*i).c_str();
+					//qDebug() << "not found " << storrage.file_path(*i).c_str();
 					filepriorities->push_back(7);
 				}
 			}
@@ -226,11 +224,11 @@ bool TorrentManager::AddTorrent(QString path, QString save_path,QMap<QString,int
 		QFile::copy(path,combine_path(QDir::currentPath().toAscii().data(),combine_path("CT_DATA",file.fileName().toAscii().data()).c_str()).c_str());
 		if (save_path_data.contains(to_hex(h.info_hash().to_string()).c_str()))
 		{
-			save_path_data[to_hex(h.info_hash().to_string()).c_str()]=save_path.toUtf8().data();
+			save_path_data[to_hex(h.info_hash().to_string()).c_str()]=save_path;
 		}
 		else
 		{
-			save_path_data.insert(to_hex(h.info_hash().to_string()).c_str(),save_path.toUtf8().data());
+			save_path_data.insert(to_hex(h.info_hash().to_string()).c_str(),save_path);
 		}
 		QFile pathDataFile("CT_DATA/path.resume");
 		if (pathDataFile.open(QFile::WriteOnly))
@@ -368,7 +366,7 @@ void TorrentManager::writeSettings()
 }
 void TorrentManager::onClose()
 {
-	////qDebug() << "start saving session";
+	//qDebug() << "start saving session";
 	writeSettings();
 	int num_outstanding_resume_data = 0;
 	std::vector<torrent_status> temp;
@@ -399,7 +397,7 @@ void TorrentManager::onClose()
 		++num_outstanding_resume_data;
 		printf("\r%d  ", num_outstanding_resume_data);
 	}
-	////qDebug("waiting for resume data %1\n", num_outstanding_resume_data);
+	//qDebug("waiting for resume data %1\n", num_outstanding_resume_data);
 
 	while (num_outstanding_resume_data > 0)
 	{
@@ -443,7 +441,7 @@ void TorrentManager::onClose()
 			save_file( combine_path("CT_DATA", to_hex(h.info_hash().to_string()) + ".resume"), out);
 		}
 	}
-	////qDebug() << "saving session state" ;
+	//qDebug() << "saving session state" ;
 	
 	{
 		entry session_state;
@@ -483,7 +481,7 @@ int TorrentManager::save_file(std::string const& filename, std::vector<char>& v)
 TorrentManager::~TorrentManager()
 {
 	
-	////qDebug() << "TorrentManager: object distruction";
+	//qDebug() << "TorrentManager: object distruction";
 	
 	onClose();
 	QApplicationSettings::FreeInstance();
@@ -496,14 +494,14 @@ TorrentManager* TorrentManager::getInstance()
 	if (_instance==NULL)
 		_instance = new TorrentManager();
 	
-	////qDebug() << "TorrentManager: someone ascked an instance";
-	////qDebug() << "TorrentManager: this is " << _instanceCount << " instance";
+	//qDebug() << "TorrentManager: someone ascked an instance";
+	//qDebug() << "TorrentManager: this is " << _instanceCount << " instance";
 	_instanceCount++;
 	return _instance;
 }
 void TorrentManager::freeInstance()
 {
-	////qDebug() << "TorrentManager: free " << _instanceCount << " instance";
+	//qDebug() << "TorrentManager: free " << _instanceCount << " instance";
 	_instanceCount--;
 	if (!_instanceCount)
 	{
@@ -573,25 +571,18 @@ opentorrent_info* TorrentManager::GetTorrentInfo(QString filename)
 }
 void TorrentManager::RemoveTorrent(torrent_handle h,bool delFiles)
 {
-	QFile::remove(combine_path("CT_DATA",h.get_torrent_info().name()+".torrent").c_str());
-	QFile::remove(combine_path("CT_DATA",to_hex(h.info_hash().to_string())+".resume").c_str());
-	save_path_data.remove(to_hex(h.info_hash().to_string()).c_str());
+
+ 	qDebug() << " removing file " << combine_path("CT_DATA",h.get_torrent_info().name()+".torrent").c_str();
+ 	if (QFile::exists(combine_path("CT_DATA",h.get_torrent_info().name()+".torrent").c_str()))
+ 	QFile::remove(combine_path("CT_DATA",h.get_torrent_info().name()+".torrent").c_str());
+ 	qDebug() << " removing file " << combine_path("CT_DATA",to_hex(h.info_hash().to_string())+".resume").c_str();
+ 	if (QFile::exists(combine_path("CT_DATA",to_hex(h.info_hash().to_string())+".resume").c_str()))
+ 	QFile::remove(combine_path("CT_DATA",to_hex(h.info_hash().to_string())+".resume").c_str());
+
+	if (save_path_data.contains(to_hex(h.info_hash().to_string()).c_str()))
+		save_path_data.remove(to_hex(h.info_hash().to_string()).c_str());
 	ses->remove_torrent(h);
-	/*if (delFiles)
-	{
-		QString path=(h.save_path()+h.name()).c_str();
-		QFileInfo info(path);
-		if (info.isDir())
-		{
-			////qDebug() << path << "is dir";
-			StaticHelpers::dellDir(path);
-		}
-		else
-		{
-			////qDebug() << path << "is file";
-			QFile::remove(path);
-		}
-	}*/
+	
 }
 QString TorrentManager::GetSessionDownloadSpeed()
 {
